+++ CorsixTH/Lua/cheats.lua
@@ -1,4 +1,5 @@
 --[[ Copyright (c) 2010 Manuel "Roujin" Wolf
+Copyright (c) 2020 lewri
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
@@ -22,146 +23,6 @@ corsixth.require("announcer")
 
 local AnnouncementPriority = _G["AnnouncementPriority"]
 
---! A dialog for activating cheats
-class "UICheats" (UIResizable)
-
----@type UICheats
-local UICheats = _G["UICheats"]
-
-local col_bg = {
-  red = 154,
-  green = 146,
-  blue = 198,
-}
-
-local col_caption = {
-  red = 174,
-  green = 166,
-  blue = 218,
-}
-
-local col_border = {
-  red = 134,
-  green = 126,
-  blue = 178,
-}
-
-local col_cheated_no = {
-  red = 36,
-  green = 154,
-  blue = 36,
-}
-
-local col_cheated_yes = {
-  red = 224,
-  green = 36,
-  blue = 36,
-}
-
---[[ Constructs the cheat dialog.
-!param ui (UI) The active ui.
-]]
-function UICheats:UICheats(ui)
-  self.cheats = ui.hospital.hosp_cheats
-  self.cheat_list = {
-    {name = "money",          func = self.cheats.cheatMoney},
-    {name = "all_research",   func = self.cheats.cheatResearch},
-    {name = "emergency",      func = self.cheats.cheatEmergency},
-    {name = "epidemic",       func = self.cheats.cheatEpidemic},
-    {name = "toggle_infected", func = self.cheats.cheatToggleInfected},
-    {name = "vip",            func = self.cheats.cheatVip},
-    {name = "earthquake",     func = self.cheats.cheatEarthquake},
-    {name = "create_patient", func = self.cheats.cheatPatient},
-    {name = "end_month",      func = self.cheats.cheatMonth},
-    {name = "end_year",       func = self.cheats.cheatYear},
-    {name = "lose_level",     func = self.cheats.cheatLose},
-    {name = "win_level",      func = self.cheats.cheatWin},
-    {name = "increase_prices", func = self.cheats.cheatIncreasePrices},
-    {name = "decrease_prices", func = self.cheats.cheatDecreasePrices},
-    {name = "decrease_prices", func = self.cheats.cheatDecreasePrices},
-  }
-
-
-  self:UIResizable(ui, 300, 200, col_bg)
-
-  self.default_button_sound = "selectx.wav"
-
-  self.modal_class = "cheats"
-  self.esc_closes = true
-  self.resizable = false
-  self:setDefaultPosition(0.2, 0.4)
-
-  local y = 10
-  self:addBevelPanel(20, y, 260, 20, col_caption):setLabel(_S.cheats_window.caption)
-    .lowered = true
-
-  y = y + 30
-  self:addColourPanel(20, y, 260, 40, col_bg.red, col_bg.green, col_bg.blue):setLabel({_S.cheats_window.warning})
-
-  y = y + 40
-  self.cheated_panel = self:addBevelPanel(20, y, 260, 18, col_cheated_no, col_border, col_border)
-
-  local function button_clicked(num)
-    return --[[persistable:cheats_button]] function(window)
-      window:buttonClicked(num)
-    end
-  end
-
-  self.item_panels = {}
-  self.item_buttons = {}
-
-  y = y + 30
-  for num = 1, #self.cheat_list do
-    self.item_panels[num] = self:addBevelPanel(20, y, 260, 20, col_bg)
-      :setLabel(_S.cheats_window.cheats[self.cheat_list[num].name])
-    self.item_buttons[num] = self.item_panels[num]:makeButton(0, 0, 260, 20, nil, button_clicked(num))
-      :setTooltip(_S.tooltip.cheats_window.cheats[self.cheat_list[num].name])
-    y = y + 20
-  end
-
-  y = y + 20
-  self:addBevelPanel(20, y, 260, 40, col_bg):setLabel(_S.cheats_window.close)
-    :makeButton(0, 0, 260, 40, nil, self.buttonBack):setTooltip(_S.tooltip.cheats_window.close)
-
-  y = y + 60
-  self:setSize(300, y)
-  self:updateCheatedStatus()
-end
-
-function UICheats:updateCheatedStatus()
-  local cheated = self.ui.hospital.cheated
-  self.cheated_panel:setLabel(cheated and _S.cheats_window.cheated.yes or _S.cheats_window.cheated.no)
-  self.cheated_panel:setColour(cheated and col_cheated_yes or col_cheated_no)
-end
-
-function UICheats:buttonClicked(num)
-  -- Only the cheats that may fail return false in that case. All others return nothing.
-  if self.cheat_list[num].func(self) ~= false then
-    if self.cheat_list[num].name ~= "lose_level" then
-      self.cheats:isCheating()
-      self:updateCheatedStatus()
-    end
-  else
-    -- It was not possible to use this cheat.
-    self.ui:addWindow(UIInformation(self.ui, {_S.information.cheat_not_possible}))
-  end
-end
-
-function UICheats:buttonBack()
-  self:close()
-end
-
-function UICheats:afterLoad(old, new)
-  if old < 141 then
-    -- Window must be closed if open for compatibility
-    local cheatWindow = self.ui:getWindow(UICheats)
-    if cheatWindow then
-      cheatWindow:close()
-    end
-  end
-  UIResizable.afterLoad(self, old, new)
-end
-
 --! A holder for all cheats in the game
 class "Cheats"
 
@@ -171,9 +32,26 @@ local Cheats = _G["Cheats"]
 -- Cheats only needs UI to function
 function Cheats:Cheats(ui)
   self.ui = ui
+  -- Cheats to appear specifically in the cheats window
+  self.cheat_list = {
+    {name = "money",          func = self.cheatMoney},
+    {name = "all_research",   func = self.cheatResearch},
+    {name = "emergency",      func = self.cheatEmergency},
+    {name = "epidemic",       func = self.cheatEpidemic},
+    {name = "toggle_infected", func = self.cheatToggleInfected},
+    {name = "vip",            func = self.cheatVip},
+    {name = "earthquake",     func = self.cheatEarthquake},
+    {name = "create_patient", func = self.cheatPatient},
+    {name = "end_month",      func = self.cheatMonth},
+    {name = "end_year",       func = self.cheatYear},
+    {name = "lose_level",     func = self.cheatLose},
+    {name = "win_level",      func = self.cheatWin},
+    {name = "increase_prices", func = self.cheatIncreasePrices},
+    {name = "decrease_prices", func = self.cheatDecreasePrices},
+  }
 end
 
-function Cheats:isCheating()
+function Cheats:announceCheat()
   local announcements = self.ui.app.world.cheat_announcements
   if announcements then
     self.ui:playAnnouncement(announcements[math.random(1, #announcements)], AnnouncementPriority.Critical)
