+++ CorsixTH/Lua/world.lua
@@ -32,6 +32,7 @@ corsixth.require("entities.humanoids.grim_reaper")
 corsixth.require("entities.humanoids.inspector")
 corsixth.require("staff_profile")
 corsixth.require("hospital")
+corsixth.require("cheats")
 corsixth.require("epidemic")
 corsixth.require("calls_dispatcher")
 corsixth.require("research_department")
